refactor(text-input): add lint rule and rename files
