+++ src/models/index.ts
@@ -1,3 +1,1 @@
 export * from './queryable-component.model';
-export * from './base-input-props.model';
-export * from './input-size.model';
