--- src/models/input-size.model.ts
@@ -1,1 +1,0 @@
-export type InputSize = 'md' | 'sm';
