+++ src/ui/text-input/text-input.component.tsx
@@ -1,34 +1,31 @@
 import React, { ComponentPropsWithRef, forwardRef } from 'react';
 import clsx from 'clsx';
 
-import { BaseInputProps } from '../../models';
-import { getTextBasedInputClasses } from '../../utils/text-based-input.utils';
+import { BaseInputProps, getTextBasedInputClasses } from '../input/';
 
 type TextInputProps = Omit<ComponentPropsWithRef<'input'>, 'size'> &
   BaseInputProps;
 
 export const TextInput = forwardRef<HTMLInputElement, TextInputProps>(
-  ({ disabled, readOnly, error, size, className, ...rest }, ref) => {
-    return (
-      <>
-        <input
-          ref={ref}
-          type="text"
-          disabled={disabled}
-          readOnly={readOnly}
-          aria-invalid={error || undefined}
-          className={clsx(
-            getTextBasedInputClasses({
-              disabled,
-              readOnly,
-              error,
-              size,
-            }),
-            className
-          )}
-          {...rest}
-        />
-      </>
-    );
-  }
+  ({ disabled, readOnly, error, size, className, ...rest }, ref) => (
+    <>
+      <input
+        ref={ref}
+        type="text"
+        disabled={disabled}
+        readOnly={readOnly}
+        aria-invalid={error || undefined}
+        className={clsx(
+          getTextBasedInputClasses({
+            disabled,
+            readOnly,
+            error,
+            size,
+          }),
+          className
+        )}
+        {...rest}
+      />
+    </>
+  )
 );
